test: add test to class aplication model gender
