--- src/main/java/com/flav/trailers/context/trailers/genders/infraestructure/persistence/mappers/MapperGender.java
@@ -1,23 +1,0 @@
-package com.flav.trailers.context.trailers.genders.infraestructure.persistence.mappers;
-
-import com.flav.trailers.context.trailers.genders.domain.models.Gender;
-import com.flav.trailers.context.trailers.genders.infraestructure.persistence.entities.GenderEntity;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.stereotype.Component;
-
-@Component
-public class MapperGender {
-
-    private final ModelMapper mapper;
-
-    public MapperGender(ModelMapper mapper) {
-        this.mapper = mapper;
-    }
-
-    public Gender toModel(GenderEntity entity) { return mapper.map(entity, Gender.class); }
-    public GenderEntity toEntity(Gender model) {
-        return  mapper.map(model, GenderEntity.class);
-    }
-
-}
