+++ src/main/resources/application-dev.properties
@@ -1,0 +1,20 @@
+server.port=5000
+
+spring.jpa.database=postgresql
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.format_sql=false
+
+spring.datasource.driver-class-name=org.postgresql.Driver
+spring.datasource.url=jdbc:postgresql://localhost/trailers
+spring.datasource.username=postgres
+spring.datasource.password=root
+
+# SWAGGER CONFIG
+springdoc.api-docs.path=/api
+springdoc.swagger-ui.path=/
+
+# FILES LOG LOCATIONS
+#logging.file.path=.
+#logging.file.name=application-logger.log
+#logging.level.root = INFO
