+++ src/main/resources/application.properties
@@ -1,11 +1,1 @@
-server.port=5000
-
-spring.jpa.database=postgresql
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.format_sql=false
-
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost/trailers
-spring.datasource.username=postgres
-spring.datasource.password=root
+spring.profiles.active =@profileActive@
